Inisialisasi projek SIMAKSIF

+++ app/Http/Controllers/SertifikatController.php
@@ -2,11 +2,11 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\Member;
 use App\Models\Sertifikat;
 use Exception;
-use GuzzleHttp\Psr7\Query;
-use Illuminate\Support\Facades\Storage;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Storage;
 use Illuminate\Support\Facades\Validator;
 
 class SertifikatController extends Controller
@@ -16,7 +16,9 @@ class SertifikatController extends Controller
         $peran_penerima = [
             'Pemateri' => 'Pemateri',
             'Peserta' => 'Peserta',
-            'Panitia' => 'Panitia'
+            'Panitia' => 'Panitia',
+            'Pengurus' => 'Pengurus',
+            'BPO' => 'Badan Pengurus Organisasi'
 
         ];
 
@@ -51,16 +53,18 @@ class SertifikatController extends Controller
         //data akan diproses di sini ketika disubmit
         $validator = Validator::make($request->all(), [
             'nomor_sertifikat' => 'required|unique:sertifikats,nomor_sertifikat',
+            'npm' => 'nullable|string|max:9',
             'nama_penerima' => 'required|string|max:100',
             'peran_penerima' => 'required',
             'nama_kegiatan' => 'required|string|max:100',
             'tanggal_sertifikat' => 'required|date',
-            'file' => 'required|file|mimes:pdf,doc,docx,jpg,jpeg,png|max:10240'
+            'file' => 'required|file|mimes:pdf,doc,docx,jpg,jpeg,png|max:2048'
         ], [
             'nomor_sertifikat.unique' => 'Nomor sertifikat harus bersifat unik.',
+            'npm.max' => 'NPM maksimal 9 karakter.',
             'nama_penerima.max' => 'Nama penerima maksimal 100 karakter.',
             'nama_kegiatan.max' => 'Nama kegiatan maksimal 100 karakter.',
-            'file_surat.max' => 'Ukuran surat maksimal 10MB.',
+            'file_surat.max' => 'Ukuran surat maksimal 2 MB.',
             'file.mimes' => 'File surat harus memiliki format pdf, doc, docx, jpg, jpeg, dan png.'
         ]);
 
@@ -75,10 +79,29 @@ class SertifikatController extends Controller
             // validate data
             $validatedData = $validator->validated();
 
+            $memberId = null;
+
+            if ($request->filled('npm')) {
+                $member = Member::where('npm', $request->npm)->first();
+
+                if (!$member) {
+                    return redirect()
+                        ->back()
+                        ->with('error', 'NPM tidak ditemukan di data anggota.')
+                        ->withInput();
+                }
+
+                $memberId = $member->id;
+            }
+
+
             $file = $request->file('file');
             $filename = now('Asia/Jakarta')->format('d-m-Y_His') . '_' . $file->getClientOriginalName();
             $file->storeAs('Sertifikat', $filename, 'public');
             $validatedData['file'] = $filename;
+
+            $validatedData['member_id'] = $memberId;
+            unset($validatedData['npm']);
 
             //simpan data
             Sertifikat::create($validatedData);
@@ -107,17 +130,19 @@ class SertifikatController extends Controller
     {
         $validator = Validator::make($request->all(), [
             'nomor_sertifikat' => 'required|unique:sertifikats,nomor_sertifikat,' . $sertifikat->id,
+            'npm' => 'nullable|string|max:9',
             'nama_penerima' => 'required|string|max:100',
             'peran_penerima' => 'required',
             'nama_kegiatan' => 'required|string|max:100',
             'tanggal_sertifikat' => 'required|date',
-            'file' => 'nullable|file|mimes:pdf,doc,docx,jpg,jpeg,png|max:10240'
+            'file' => 'nullable|file|mimes:pdf,doc,docx,jpg,jpeg,png|max:2048'
 
         ], [
             'nomor_sertifikat.unique' => 'Nomor sertifikat harus bersifat unik.',
+            'npm.max' => 'NPM maksimal 9 karakter.',
             'nama_penerima.max' => 'Nama penerima maksimal 100 karakter.',
             'nama_kegiatan.max' => 'Nama kegiatan maksimal 100 karakter.',
-            'file_surat.max' => 'Ukuran surat maksimal 10MB.',
+            'file_surat.max' => 'Ukuran surat maksimal 2 MB.',
             'file.mimes' => 'File surat harus memiliki format pdf, doc, docx, jpg, jpeg, dan png.'
         ]);
 
@@ -131,30 +156,51 @@ class SertifikatController extends Controller
         try {
             //function yang memproses saat update disubmit
             // validate data
-            $validatedData = $validator->validate();
-            $isChanged = false;
+            $validatedData = $validator->validated();
+            $memberId = null;
 
-            $mainFields = ['nomor_sertifikat', 'nama_penerima', 'peran_penerima', 'nama_kegiatan', 'tanggal_sertifikat'];
-            foreach ($mainFields as $field) {
-                if ($request->filled($field) && $sertifikat->$field != $request->$field) {
-                    $isChanged = true;
-                    break;
+            if ($request->filled('npm')) {
+                $member = Member::where('npm', trim($request->npm))->first();
+
+                if (!$member) {
+                    return redirect()
+                        ->back()
+                        ->with('error', 'NPM tidak ditemukan di data anggota.')
+                        ->withInput();
                 }
+
+                $memberId = $member->id;
             }
 
+            $validatedData['member_id'] = $memberId;
+            unset($validatedData['npm']);
+
+            $sertifikat->fill($validatedData);
             if ($request->hasFile('file')) {
-                if ($sertifikat->file) {
+
+                // hapus file lama
+                if ($sertifikat->file && Storage::disk('public')->exists('Sertifikat/' . $sertifikat->file)) {
                     Storage::disk('public')->delete('Sertifikat/' . $sertifikat->file);
                 }
+
                 $file = $request->file('file');
                 $filename = now('Asia/Jakarta')->format('d-m-Y_His') . '_' . $file->getClientOriginalName();
                 $file->storeAs('Sertifikat', $filename, 'public');
-                $validatedData['file'] = $filename;
-                $isChanged = true;
+
+                $sertifikat->file = $filename;
             }
 
-            // update data
-            $sertifikat->update($validatedData);
+            if ($sertifikat->isDirty()) {
+                $sertifikat->save();
+
+                return redirect()
+                    ->route('sertifikat.index')
+                    ->with('success', 'Data berhasil diperbarui.');
+            }
+
+            return redirect()
+                ->route('sertifikat.index')
+                ->with('info', 'Tidak ada perubahan data.');
 
             // redirect ke index ketika berhasil disimpan
             if ($isChanged) {
@@ -163,7 +209,7 @@ class SertifikatController extends Controller
             return redirect()->route(route: 'sertifikat.index')->with('info', 'Tidak ada perubahan data.');
         } catch (\Exception $e) {
             return redirect()->route('sertifikat.index')
-            ->with('error', 'Gagal memperbarui data. Silahkan coba lagi. ');
+                ->with('error', 'Gagal memperbarui data. Silahkan coba lagi. ');
         }
     }
